power: qpnp-charger: Port Chrg OCV
This change ports the following CRs to Falcon
(CR) and (CR)

This will add the Charge OCV calculation
which will prevent charging for 4 minutes
on charger Insertion so that an OCV can be
calculated.  This iwll only happen if SOC
is between 20% and 60% and it has been 1
week since the last calculation.  There can
also only be 1 attempt every hour.

Change-Id: Idd430fd004e8a17555f8b34d6d030777d4145d87
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>