power: qpnp-charger: add OCP support for SMBBP/SMBCL
Add over current protection support for the USB
peripheral of SMBBP and SMBCL devices.

The OCP interrupt is used to notify and clear
of over current conditions when reverse boosting for
USB OTG. In an over current condition the hardware
opens the OTG FET but leaves the boost on. The
interrupt handler clears the OCP LATCH and turns
back on the OTG switch.

Change-Id: I78ad69d5b17bb7f9499b11f3cbde335e0e29ec79
Signed-off-by: David Keitel <dkeitel@codeaurora.org>