power: qpnp-charger: Handle missed disconnects in heartbeat
The update heartbeat check in the qpnp-charger driver ignores cases where the
usb bus is suspended. Hence the workaround for missed USB-IN irq's does not
execute for cases where the bus is suspended and the USB-IN irq is not recieved.
Modify the workaround by checking the USB_VALID bit against the chip->usb_present
which is always updated in the USB-IN irq handler.

Change-Id: Iadd0953cfdce3b09b0cc0f6a201e2de2018048da
Signed-off-by: Ravi Vembu <a18394@motorola.com>
Reviewed-on: http://gerrit.mot.com/629124
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Kenneth Kessler <kennykessler@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
Reviewed-by: Vidyadhar Gurram <rnh786@motorola.com>
Reviewed-by: Kumar Venkatesh <kumarv@motorola.com>