ASoC: msm8226: set value for SPK_RCV_SEL always
Titan uses top speaker as earpiece. SPK_RCV_SEL is used to select
the correct mode (ie either speaker or earpiece).

The defaul value is assumed to be 0 ie speaker. However, as
SPK_RCV_SEL is a switch, assuming that it will be in default
mode is erroneous. It needs be set to the correct mode evertime
so as to match the accessory.

Change-Id: I9262cd999934b25fc61e3c19db6a0c25f107426f
Signed-off-by: Rajasekhar Goparaju <chb764@motorola.com>
Reviewed-on: http://gerrit.mot.com/629760
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Vladimir Karpovich <vkarpovich@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>